requirements.txt
asgiref==3.5.2
certifi==2022.12.07
dj-database-url==1.2.0
Django==4.1
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
gunicorn==20.1.0
Markdown==3.4.1
mysqlclient==2.1.1
Pillow==9.4.0
psycopg2==2.9.5
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2021.1
whitenoise==6.3.0
wincertstore==0.2
